[#115] Refactor SurveyDetailScreen to preview
