+++ app/src/main/java/com/kks/nimblesurveyjetpackcompose/ui/presentation/survey/SurveyDetailStartScreen.kt
@@ -1,0 +1,53 @@
+package com.kks.nimblesurveyjetpackcompose.ui.presentation.survey
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.kks.nimblesurveyjetpackcompose.model.Survey
+import com.kks.nimblesurveyjetpackcompose.ui.presentation.common.SurveyBoldText
+import com.kks.nimblesurveyjetpackcompose.ui.theme.NeuzeitFamily
+import com.kks.nimblesurveyjetpackcompose.ui.theme.White70
+
+@Composable
+fun SurveyDetailStartScreen(survey: Survey, modifier: Modifier = Modifier) {
+    Column(modifier = modifier, verticalArrangement = Arrangement.spacedBy(16.dp)) {
+        SurveyBoldText(
+            text = survey.title,
+            fontSize = 28.sp,
+            modifier = Modifier
+                .padding(horizontal = 20.dp)
+                .fillMaxWidth()
+        )
+        Text(
+            text = survey.description,
+            fontFamily = NeuzeitFamily,
+            fontWeight = FontWeight.Normal,
+            color = White70,
+            fontSize = 17.sp,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 20.dp)
+        )
+    }
+}
+
+@Preview(showBackground = true, backgroundColor = 0)
+@Composable
+fun SurveyDetailStartScreenPreview() {
+    SurveyDetailStartScreen(
+        survey = Survey(
+            id = "",
+            coverImagePlaceholderUrl = "https://dhdbhh0jsld0o.cloudfront.net/m/c96c480fc8b69d50e75a_",
+            title = "Tree Tops Australia",
+            description = "We'd love to hear from you!"
+        )
+    )
+}
